backend/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Cohere AI
cohere>=5.0.0

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.26.0
